+++ plone/schemaeditor/tests/fixtures.py
@@ -32,7 +32,7 @@ def log_event(object, event):
     print('[event: {0} on {1}]'.format(
         event.__class__.__name__,
         object.__class__.__name__,
-    ))
+    ))       # noqa
 
 
 class EditForm(EditForm):
@@ -43,6 +43,7 @@ class EditForm(EditForm):
     def update(self):
         self.fields = field.Fields(IDummySchema)
         super(EditForm, self).update()
+
 
 EditView = layout.wrap_form(EditForm)